JwtService refactoring
